禁卡表的类型修复：搜索

+++ mobile/src/main/java/cn/garymb/ygomobile/loader/CardSearchInfo.java
@@ -15,7 +15,7 @@ class CardSearchInfo {
     long race, category;
     String atk, def;
     boolean islink;
-    List<Long> inCards;
+    List<Integer> inCards;
     long[] types;
     long setcode;
